Update 3 - Base Conversion
Out of playful curiosity I added a well-known function for base
conversion and added a terminal command to convert the answer of  a
mathematical expression to base12.
I intend to eventually do the calculation in base12 and change the
output text such that it's less confusing to base10 natives.
[Substituting letters for numbers], but that was a bit much for me to
work on today.