Update 4 - Bookmarks
With this update I added a new bookmark feature to the browser.
Changelog:
- Added function rezBookmark which takes the arguments url & imgsrc.
- Added function addBookmarks that reads a prompt-derived string for url
& imgsrc data. Each link is seperated by whitespace, and the two
variables are seperated by commas.
- Added function goto which is called by newly created image links to
dynamically change the iframe-browser's src property.
- Modified the rezBrowser function to place the addressBar, goButton,
and new bookmarkButton in a div called "toolBox" that is now appended to
the browser wrapper before the browserBox div.
- Included a new icon named test.jpg that is the default for new
bookmarks.
BUGS:
-New bookmarks are supposed to be appended to the toolBox, however, they
seem destined to be appended to the main menu.
This is very minor and so I decided to include it in this push.