Update 6 - Self-Removing Elements
Generally made most Rez Commands capable of DeRezzing if an element
already exists. One Rez command in particular was left out:
"rezBrowserTo", this is because it will operate differently, if an
existing browser is detected it will [in the future] change the .src
property of the iframe. I may use a boolean argument to toggle the
functionality. [someone may want multiple windows]
- Added a new icon for opening the browser on the main menu, no more
need for the command line to do so.
- Terminal now closes by clicking the icon used to open it.
- Browser now closes by clicking the icon used to open it.
- The AboutUs div now closes by clicking the icon used to open it.
- Changed the DwitekLogo variable to DwitekIcon to fit the naming style.
- The container variable in the rezCommandLine() function is now global
to the function because it is used in both the If and Else statements.