Update 7 - Local Storage!
I changed a few functions to make this update happen, and it got a
little frustrating to be honest!
With a little logic, however, all problems are overcome.

Changelog:
-Revised the Bookmark Creation, a new function "inputBookmarks()"
prompts now instead of the addBookmarks function, inputBookmarks also
incorporates the novel feature of this update: it sets the localStorage
variable "linkList" to the prompt's value if the variable does not
exist, otherwise it adds the old input to the new.

- addBookmarks is now passed the input from inputBookmarks function.

-Added function clearBookmarks() which clears the local storage,
accessible through the terminal/commandLine.

-Start() function now checks for the localStorage variable and appends
links by calling addBookmarks, passing it the variable from
localStorage.

-Edited to help command to include an updated list of commands.