Update 9 - Flexibility
I looked at my function for base12Calculation and decided it made much
more sense to make the function capable of taking a mathematical
expression and rendering it in any base. 14 lines of code become 5.

-The new command line program "evalBase" is used as such: "evalBase
12*12*12 20".

You call the program, the program searches through the terminal's array
for the variables [mathematical expression] && [The Base to render the
answer in].