Update 10 - Functions are Objects!
The more I write the more I see.

- New Function returnIcon() takes a string descriptor as it's argument
and returns the corresponding icon object.

- Start function now calls a loop that iterates through an array of
Program descriptors stored as strings, which are passed to the
returnIcon() object which returns the corresponding icon objects which
are placed in the array which are then rezzed on the page from the
array.

- dwitekStaff array now holds my image object/link returned from the
returnIcon() object which is then rezzed.