Update 20 - Tie-ins
Tied in the game to a terminal call so that one can open the game from
the wrapper.

Changelog:
- In the last few updates I pushed I experimented with:

)Adsense functionality and manipulation of Google's adsense element to
make it useful for web-applications.

)I designed a small height-based terrain generation and rendering
program I'm calling 'gridPlexa'.
_At this point in time it is very simple [and buggy], but it shows that
game development with Javascript and HTML can be accomplished by
complete amateurs without the use of extensive libraries or the canvas
element.